+++ src/blob/index/bptree/core.rs
@@ -51,9 +51,10 @@ where
         headers: &InMemoryIndex<K>,
         meta: Vec<u8>,
         recreate_index_file: bool,
+        blob_size: u64,
     ) -> Result<Self> {
         clean_file(path, recreate_index_file)?;
-        let res = Self::serialize(headers, meta)?;
+        let res = Self::serialize(headers, meta, blob_size)?;
         let (mut header, metadata, buf) = res;
         let file = File::create(path, ioring)
             .await
@@ -110,9 +111,9 @@ where
         self.read_headers(leaf_offset, key, &mut buf).await
     }
 
-    async fn get_records_headers(&self) -> Result<(InMemoryIndex<K>, usize)> {
+    async fn get_records_headers(&self, blob_size: u64) -> Result<(InMemoryIndex<K>, usize)> {
         let mut buf = self.file.read_all().await?;
-        self.validate_header(&mut buf).await?;
+        self.validate_header(&mut buf, blob_size).await?;
         let offset = self.metadata.leaves_offset as usize;
         let records_end = FileIndexTrait::<K>::file_size(self) as usize;
         let records_buf = &buf[offset..records_end];
@@ -140,7 +141,7 @@ where
         self.read_header(leaf_offset, key, &mut buf).await
     }
 
-    fn validate(&self) -> Result<()> {
+    fn validate(&self, blob_size: u64) -> Result<()> {
         // FIXME: check hash here?
         if !self.header.is_written() {
             let param = ValidationErrorKind::IndexIsWritten;
@@ -151,6 +152,18 @@ where
         if self.header.version() != HEADER_VERSION {
             let param = ValidationErrorKind::IndexVersion;
             return Err(Error::validation(param, "Index Header version is not valid").into());
+        }
+        if self.header.blob_size() != blob_size {
+            let param = ValidationErrorKind::IndexBlobSize;
+            return Err(Error::validation(
+                param,
+                format!(
+                    "Index Header is for blob of size {}, but actual blob size is {}",
+                    self.header.blob_size(),
+                    blob_size
+                ),
+            )
+            .into());
         }
         if self.header.magic_byte() != INDEX_HEADER_MAGIC_BYTE {
             let param = ValidationErrorKind::IndexMagicByte;
@@ -346,8 +359,8 @@ where
         Ok(())
     }
 
-    async fn validate_header(&self, buf: &mut Vec<u8>) -> Result<()> {
-        self.validate()?;
+    async fn validate_header(&self, buf: &mut Vec<u8>, blob_size: u64) -> Result<()> {
+        self.validate(blob_size)?;
         if !Self::hash_valid(&self.header, buf)? {
             let param = ValidationErrorKind::IndexChecksum;
             return Err(Error::validation(param, "header hash mismatch").into());
@@ -392,9 +405,10 @@ where
     fn serialize(
         headers_btree: &InMemoryIndex<K>,
         meta: Vec<u8>,
+        blob_size: u64,
     ) -> Result<(IndexHeader, TreeMeta, Vec<u8>)> {
         Serializer::new(headers_btree)
-            .header_stage(meta)?
+            .header_stage(meta, blob_size)?
             .tree_stage()?
             .build()
     }