+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
+- Add blob size to index header (#153)
 - Add magic byte to index header (#152)
 
 #### Changed